requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
black==24.4.2
CacheControl==0.14.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-ckeditor==6.7.2
django-debug-toolbar==4.4.2
django-js-asset==3.0.1
django-unfold==0.63.0
djangorestframework==3.15.1
django-filter==24.1
factory_boy==3.3.3
Faker==37.0.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.8
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pep517==0.13.1
pillow==10.3.0
pipdeptree==2.24.0
platformdirs==4.2.2
pluggy==1.5.0
progress==1.6
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.1.1
pytoml==0.1.21
pytz==2024.1
requests==2.32.2
retrying==1.3.4
selenium==4.30.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.6.0
wsproto==1.2.0
